brain_tumor_app_deployment/deployment/requirements.txt
flask==3.1.0
flask-login==0.6.3
tensorflow==2.19.0
scikit-image==0.25.2
scikit-learn==1.6.1
lime==0.2.0.1
shap==0.47.1
matplotlib==3.10.1
pillow==10.2.0
opencv-python-headless==4.11.0.86
numpy==2.1.3
